+++ examples/multiple_functions/app.py
@@ -1,0 +1,6 @@
+from scw_serverless.app import Serverless
+
+app = Serverless("multiple-functions")
+
+import first  # pylint: disable=all # noqa
+import second  # noqa
